evm: fix test_transceiverIncompatibleNttManager
This test started failing as of forge v1.3.1, specifically due to this
PR https://github.com/foundry-rs/foundry/pull/10446.

The PR implements better error reporting *when the tests run in verbose
mode*, so the test that previously expected an empty revert string now
fails in verbose mode, because forge now returns an error about calling
a contract with no bytecode.

Actually this test was supposed to test (according to the comment) that
the transceiver constructor revert if the manager doesn't have the token
method, but that's not what it was testing. So we change the test to
actually use a dummy contract that has no token method. This returns an
empty revert string, so everything is good. Until forge includes an
override for that too at least