cli: support deploying different EVM variants
We have a growing list of NTT manager variants (the standard, one
without rate limiting, one that wraps WETH, and maybe we'll have more)

To achieve this, we have to do a couple things.

Firstly, we use deploy script in the worktree by default.

The current (as in before this commit) method of overriding the deploy
script with the one bundled into the cli was introduced at the very
beginning to allow deploying versions that didn't have the updated
deploy script available. This was actually already redundant by the time
the first NTT release was cut.  It's very rigid behaviour because it
doesn't allow introducing breaking changes to the deploy script that
don't work with older versions of the contract.

So: we simply use the deploy script that's in the work tree.

This allows us to evolve the deploy script over time, such as in the
current case where we add support for the different variants.

One problem is that prior to this commit, the deploy scripts at various
version releases don't work (for example no currently released version
supports non-0 wormhole fees.). So, if the deploy script in the worktree
is old (doesn't contain the newly introduced DEPLOY_SCRIPT_VERSION
comment), then we override it with a known version that works. Otherwise
we just use the worktree version, which will work going forward.